Move dataclasses to models.py
Move dataclasses used for representing info about PR comments to
models.py so that all classes for storing data are placed together.
